Refactor problem panel
